:recycle: refactor: atoms 파일들 stores에서 atoms로 이관
