+++ src/atoms/alert.tsx
@@ -11,3 +11,13 @@ export const alertAtom = atom<AlertStateType>({
   props: {},
   responseHandler: () => null,
 });
+
+export const closeAlertAtom = atom(null, (get, set, state: boolean) => {
+  const { responseHandler } = get(alertAtom);
+  responseHandler(state);
+  set(alertAtom, {
+    Component: () => <div />,
+    props: {},
+    responseHandler: () => {},
+  });
+});
